% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeexplode.R
\name{r_read_as_rmd_table}
\alias{r_read_as_rmd_table}
\title{Title}
\usage{
r_read_as_rmd_table(r_script_path = "docs/r_script_test.R")
}
\arguments{
\item{r_script_path}{}
}
\value{

}
\description{
Title
}
