% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeexplode.R
\name{rmd_code_explode}
\alias{rmd_code_explode}
\title{Rmarkdown file as input, returns rmd prepped to flipbook}
\usage{
rmd_code_explode(
  rmd_path = "README.Rmd",
  rmd_output = stringr::str_replace(rmd_path, "\\\\.Rmd|\\\\.rmd", "_flipbook.Rmd"),
  render = TRUE
)
}
\arguments{
\item{rmd_path}{a file path, character string}

\item{rmd_output}{a file path, where to save rmd}

\item{render}{if true, render}
}
\value{

}
\description{
Rmarkdown file as input, returns rmd prepped to flipbook
}
