% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeexplode.R
\name{r_code_explode}
\alias{r_code_explode}
\title{Title}
\usage{
r_code_explode(
  r_script_path = "docs/r_script_test.R",
  rmd_output = stringr::str_replace(rmd_path, "\\\\.R$|\\\\.r$", "_flipbook.Rmd"),
  render = TRUE,
  font_size = "70\%"
)
}
\arguments{
\item{r_script_path}{path to .Rmd script}

\item{rmd_output}{a file path, where to save rmd}

\item{render}{if true, render}
}
\value{

}
\description{
Title
}
