% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeexplode.R
\name{rmd_code_explode}
\alias{rmd_code_explode}
\title{Title}
\usage{
rmd_code_explode(
  rmd_path = "README.Rmd",
  rmd_output = "explodedcode.Rmd",
  render = F
)
}
\arguments{
\item{render}{}
}
\value{

}
\description{
Title
}
